Rename: 컴포넌트 확장에 따른 디렉토리 수정 (CreatePost -> AddOrEditPost)
